% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title CLT Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
A histogram of the sample poisson distribution with a denisty curve, a barplot, and a probability function.
}
\description{
Title CLT Poisson Distribution
}
\examples{
\dontrun {mycltp=function(n,iter,lambda)}
}
