% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyNCurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve function for Lab 6}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{a}{}
}
\value{
Plots the curve, shades the probability, and calculates the area.
}
\description{
Curve function for Lab 6
}
\examples{

}
