% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{My sample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Iteration}

\item{time}{Time delay between executions}
}
\value{
A barplot
}
\description{
My sample
}
\examples{
\dontrun{a <- 1000, b <- 10, c <- 1; mysample(n=a,iter=b,time=c)}
}
